backend/backend/requirements.txt
annotated-types
anyio
bcrypt
blis
catalogue
certifi
cffi
charset-normalizer
click
cloudpathlib
colorama
confection
cryptography
cymem
dnspython
ecdsa
email_validator
fastapi
fastapi-cli
fastapi-cloud-cli
greenlet
h11
httpcore
httptools
httpx
idna
itsdangerous
Jinja2
langcodes
language_data
marisa-trie
markdown-it-py
MarkupSafe
mdurl
murmurhash
numpy
orjson
packaging
preshed
psycopg2-binary
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0.tar.gz#sha256=b86d2fcb022bbe1dffa8692852c9094591e93d0299b65c5bdf92f7940ae2c47c
pyasn1
pycparser
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
Pygments
python-dotenv
python-jose
python-multipart
PyYAML
requests
rich
rich-toolkit
rignore
rsa
sentry-sdk
setuptools
shellingham
six
smart_open
sniffio
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
srsly
starlette
thinc
tqdm
typer
typing-inspection
typing_extensions
ujson
urllib3
uvicorn
wasabi
watchfiles
weasel
websockets
wrapt


